[XLA] Add some ReduceWindow tests and move IsTrivialWindowDimension into window_util.
PiperOrigin-RevId: 237906295
